requirements.txt
Flask==2.3.3
biopython==1.83
